수정,,제발,,

+++ sparta_news/articles/views.py
@@ -1,4 +1,3 @@
-
 from django.shortcuts import render, get_object_or_404
 from rest_framework import generics
 from rest_framework.response import Response
@@ -14,18 +13,15 @@ from rest_framework_simplejwt.tokens import AccessToken
 
 
 class CommentListCreatView(generics.ListAPIView):
-
-
     # permission_classes = [IsAuthenticatedOrReadOnly]
     serializer_class = CommentSerializer
 
-    def get(self,request,article_id):
+    def get(self, request, article_id):
         comments = Comments.objects.filter(article_id=article_id)
-        serializer = CommentSerializer(comments,many=True)
+        serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data)
 
-
-    def post(self,request,article_id):
+    def post(self, request, article_id):
         # user = request.user
         user = get_object_or_404(Accounts, pk=1)
         article = get_object_or_404(Articles, pk=article_id)
@@ -33,65 +29,61 @@ class CommentListCreatView(generics.ListAPIView):
         content = request.data.get("content")
 
         if not content:
-            return Response({"error":"댓글 내용 입력이 없습니다"})
-        
+            return Response({"error": "댓글 내용 입력이 없습니다"})
+
         if len(content) < 10:
-            return Response({"error":"댓글의 내용을 10글자 이상 작성해주세요"})
-        
-        comment = Comments.objects.create(content=content,user=user,article=article)
+            return Response({"error": "댓글의 내용을 10글자 이상 작성해주세요"})
+
+        comment = Comments.objects.create(content=content, user=user, article=article)
         serializer = CommentSerializer(comment)
-        return Response(serializer.data,status=status.HTTP_201_CREATED)
-    
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
 
 class CommentDetailView(generics.ListAPIView):
-    
     """
     HTTP 프로토콜 수정 제안 사항
-    
+
     1. 보통 새 데이터의 추가는 POST()를 사용하고 - 블로그 글을 포스팅한다 -,
-    기존에 존재하는 데이터의 수정은 PUT()을 사용합니다. 
-    
+    기존에 존재하는 데이터의 수정은 PUT()을 사용합니다.
+
     PUT()이 댓글 수정인지 확실하지가 않은데, 만약 새 댓글을 작성하는 것이라면
     API 사용 기준을 따르려면 PUT이 아닌 POST를 사용하는 것이 좋을 듯 합니다. :)
-    
+
     2. user 조회하는 코드의 pk 값도 request.user의 id값을 참조하도록 수정부탁드립니다. :)
     """
-       
-    
+
     # permission_classes = [IsAuthenticated]
     serializer_class = CommentSerializer
 
-    def put(self,request,comment_id):
+    def put(self, request, comment_id):
         user = get_object_or_404(Accounts, pk=1)
-        comment = get_object_or_404(Comments,pk=comment_id)
+        comment = get_object_or_404(Comments, pk=comment_id)
         content = request.data.get("content")
         if not content:
-            return Response({"error":"댓글 내용 입력이 없습니다"})
-        
+            return Response({"error": "댓글 내용 입력이 없습니다"})
+
         if len(content) < 5:
-            return Response({"error":"댓글의 내용을 5글자 이상 작성해주세요"})
-        
-        #if request.user == comment.user:
+            return Response({"error": "댓글의 내용을 5글자 이상 작성해주세요"})
+
+        # if request.user == comment.user:
         if user == comment.user:
             comment.content = content
             comment.save()
             serializer = CommentSerializer(comment)
             return Response(serializer.data)
-        return Response({"error":"권한 없는 사용자입니다."},status=status.HTTP_403_FORBIDDEN)
+        return Response({"error": "권한 없는 사용자입니다."}, status=status.HTTP_403_FORBIDDEN)
 
-
-    def delete(self,request,comment_id):
+    def delete(self, request, comment_id):
         user = get_object_or_404(Accounts, pk=1)
         comment = get_object_or_404(Comments, pk=comment_id)
-        #if request.user == comment.user:
+        # if request.user == comment.user:
         if user == comment.user:
             comment.delete()
-            return Response({"message":"댓글이 삭제 되었습니다."},status=status.HTTP_204_NO_CONTENT)
-        return Response({"error":"권한 없는 사용자입니다."},status=status.HTTP_403_FORBIDDEN)
+            return Response({"message": "댓글이 삭제 되었습니다."}, status=status.HTTP_204_NO_CONTENT)
+        return Response({"error": "권한 없는 사용자입니다."}, status=status.HTTP_403_FORBIDDEN)
 
 
 class ArticleListView(APIView):
-
     """
     확인 필요 사항
 
@@ -99,7 +91,7 @@ class ArticleListView(APIView):
     확인을 한 번 해봐야 할 듯 합니다.
 
     self.permission_classes = [IsAuthenticated]로 인해서
-    클래스 상단에 선언한 permission_classes값이 [IsAuthenticatedOrReadOnly]가 
+    클래스 상단에 선언한 permission_classes값이 [IsAuthenticatedOrReadOnly]가
     [IsAuthenticated]로 변경되기 때문입니다. :)
     """
 
@@ -118,10 +110,9 @@ class ArticleListView(APIView):
         if serializer.is_valid(raise_exception=True):
             serializer.save(user=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-          
-          
-class ArticlesDetailAPIView(APIView):
 
+
+class ArticlesDetailAPIView(APIView):
     """
     반드시 수정이 필요한 부분에 대한 요청
 
@@ -131,7 +122,6 @@ class ArticlesDetailAPIView(APIView):
 
     2. 계정 테스트로 넣어두신 주석처리된 get()은 삭제하셔도 됩니다 :)
     """
-
 
     def get_object(self, pk):
         try:
@@ -144,24 +134,27 @@ class ArticlesDetailAPIView(APIView):
         get_obj = self.get_object(pk)
         serializer = ArticlesSerializer(get_obj)
         return Response(serializer.data)
-      
-#    def get(self, request):
-#        user = authenticate(username=request.data['username'], password=request.data['password'])
-#        if user is not None:
-#            access_token = AccessToken.for_user(user)
-#            return Response({'access_token': str(access_token)}, status=200)
-#        else:
-#            return Response({'error': 'Invalid credentials'}, status=400)
 
     def put(self, request, pk):
         article = self.get_object(pk=pk)
+
+        if not request.user == article.author:
+            return Response({"error": "권한 없는 사용자입니다."}, status=status.HTTP_403_FORBIDDEN)
+
         serializer = ArticlesSerializer(article, data=request.data, partial=True)
+
         if serializer.is_valid(raise_exception=True):
-            serializer.save()
+            serializer.save()  # 유효한 경우 저장
             return Response(serializer.data, status=200)
+        else:
+            # 유효하지 않은 경우 에러 응답
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
         article = self.get_object(pk=pk)
+        if not request.user == article.author:
+            return Response({"error": "권한 없는 사용자입니다."}, status=status.HTTP_403_FORBIDDEN)
+
         article.delete()
         data = {"pk": f"{pk} is deleted."}
         return Response(data, status=200)
